Feature/endpoints atas

+++ src/main/java/com/br/zup/gerenciadordeguildas/repositories/AtaRepository.java
@@ -3,5 +3,5 @@ package com.br.zup.gerenciadordeguildas.repositories;
 import com.br.zup.gerenciadordeguildas.entities.Ata;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AtaRepository extends CrudRepository<Ata, Long> {
+public interface AtaRepository extends CrudRepository<Ata, Integer> {
 }
